smart-hire-main (2)/smart-hire-main/smart-hire-main/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
reportlab==4.2.2
python-docx==1.1.2
PyPDF2==3.0.1
PyMuPDF==1.24.9
python-dateutil==2.9.0.post0
numpy==2.0.1
scipy==1.14.1
nltk==3.9.1
textstat==0.7.3
scikit-learn==1.5.2
language-tool-python==2.9.0
gunicorn==22.0.0
